requirements.txt
googlemaps>=4.10.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
rich>=13.7.0
Pygments>=2.17.0
langflow>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
